[ET-VK][Op Redesign][6/n] Merge ArithmeticPrepack into PrepackNode
There's a lot of shared logic between
- `add_staging_to_tensor_node()`, which handles I/O data on execute(), and
- `ArithmeticPrepack`'s simple prepacking, on prepack().

Both just copy data to and from GPU, without any manipulation. Hence, I've decided to consolidate shared logic in this diff as well. Here are the final results:
+ Make `PrepackNode` a final class.
+ Remove all references of `impl/Packing.h`.
- Extract shared util functions to new `StagingUtils.h/cpp`.

Differential Revision: [D54504449](https://our.internmc.facebook.com/intern/diff/D54504449/)

[ghstack-poisoned]
