+++ backends/vulkan/runtime/graph/ops/impl/Staging.h
@@ -18,69 +18,6 @@ namespace at {
 namespace native {
 namespace vulkan {
 
-//
-// Functions to memcpy data into staging buffer
-//
-
-void memcpy_to_mapping(
-    const void* src,
-    api::MemoryMap& dst_mapping,
-    const size_t nbytes,
-    const api::ScalarType dtype);
-void memcpy_from_mapping(
-    const api::MemoryMap& src_mapping,
-    void* dst,
-    const size_t nbytes,
-    const api::ScalarType dtype);
-
-//
-// Utility functions for memcpy
-//
-
-template <typename T>
-void memcpy_to_mapping_impl(
-    const void* src,
-    api::MemoryMap& dst_mapping,
-    const size_t nbytes) {
-  T* data_ptr = dst_mapping.template data<T>();
-  memcpy(data_ptr, reinterpret_cast<const T*>(src), nbytes);
-}
-
-template <typename T>
-void memcpy_from_mapping_impl(
-    api::MemoryMap& src_mapping,
-    void* dst,
-    const size_t nbytes) {
-  T* data_ptr = src_mapping.template data<T>();
-  memcpy(reinterpret_cast<T*>(dst), data_ptr, nbytes);
-}
-
-//
-// Functions to copy data into and out of a staging buffer
-//
-
-void copy_ptr_to_staging(
-    const void* src,
-    api::StorageBuffer& staging,
-    const size_t nbytes);
-void copy_staging_to_ptr(
-    api::StorageBuffer& staging,
-    void* dst,
-    const size_t nbytes);
-
-//
-// Functions to record copying data between a staging buffer and a vTensor
-//
-
-void encode_copy_to_vtensor(
-    api::Context* context,
-    api::StorageBuffer& staging,
-    vTensor& tensor);
-
-//
-// Functions to initialize ExecuteNode
-//
-
 void add_staging_to_tensor_node(
     ComputeGraph& graph,
     const ValueRef in_staging,
@@ -90,12 +27,13 @@ void add_tensor_to_staging_node(
     const ValueRef in_tensor,
     const ValueRef out_staging);
 
-//
-// Functions to get shaders
-//
+struct StagingParams final {
+  api::utils::ivec3 extents;
+  int32_t plane_size;
+  api::utils::ivec2 channel_info;
+};
 
-api::ShaderInfo get_nchw_to_image_shader(const vTensor& v_dst);
-api::ShaderInfo get_image_to_nchw_shader(const vTensor& v_src);
+ValueRef prepack_if_tensor_ref(ComputeGraph& graph, const ValueRef v);
 
 } // namespace vulkan
 } // namespace native
